phase_2/problem_2/requirements.txt
category_encoders==2.6.1
evidently==0.4.0
fastapi==0.100.0
imbalanced_learn==0.10.1
imblearn==0.0
pandas==2.0.1
pydantic==1.10.7
PyYAML==6.0
scikit_learn==1.2.2
scipy==1.10.1
uvicorn==0.22.0
xgboost==1.7.5
pyarrow==11.0.0
fastparquet
